requirement.txt
# ============================================
# WEB SCRAPING DEPENDENCIES
# ============================================
playwright==1.40.0
beautifulsoup4==4.12.2
aiohttp==3.9.1

# ============================================
# DATA PROCESSING & ANALYSIS
# ============================================
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# ============================================
# DATA VALIDATION & MODELS
# ============================================
pydantic==2.5.3
pydantic-settings==2.1.0

# ============================================
# DATABASE
# ============================================
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# ============================================
# API FRAMEWORK (for later phases)
# ============================================
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# ============================================
# ML & AI (Phase 2+)
# ============================================
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
shap==0.44.0

# LLM & Vector Store (Phase 3+)
openai==1.6.1
anthropic==0.8.1
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.22
tiktoken==0.5.2

# ============================================
# MLOPS & MONITORING (Phase 4+)
# ============================================
mlflow==2.9.2
prometheus-client==0.19.0

# ============================================
# UTILITIES
# ============================================
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
tenacity==8.2.3
httpx==0.26.0

# ============================================
# TESTING
# ============================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==22.0.0

# ============================================
# CODE QUALITY
# ============================================
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# ============================================
# DEVELOPMENT TOOLS
# ============================================
jupyter==1.0.0
ipython==8.19.0
notebook==7.0.6